+++ backend/src/main/kotlin/backend/api/App.kt
@@ -1,7 +1,7 @@
 /*
  * This Kotlin source file was generated by the Gradle 'init' task.
  */
-package api
+package backend.api
 
 import io.javalin.Javalin
 
@@ -9,6 +9,4 @@ fun main() {
     val app = Javalin.create { config -> config.enableWebjars() }.start(7070)
     apiRoutes(app)
     app.get("/") { ctx -> ctx.result("Hello World") }
-
-
 }
