+++ backend/src/main/kotlin/backend/models/order/Dough.kt
@@ -1,9 +1,9 @@
-package models.order
+package backend.models.order
 
 /**
  * Lets assume Italian style dough is worth as much as American :)
  */
-enum class Dough{
+enum class Dough {
     ITALIAN,
     AMERICAN
 }
