+++ backend/src/main/kotlin/backend/models/order/Order.kt
@@ -1,0 +1,5 @@
+package backend.models.order
+
+import backend.models.pizza.Pizza
+
+class Order(val size: Size, val dough: Dough, pizza: Pizza)
