+++ backend/src/main/kotlin/backend/models/order/Size.kt
@@ -1,4 +1,4 @@
-package models.order
+package backend.models.order
 
 enum class Size(val value: Int) {
     SMALL(30), MEDIUM(70), LARGE(120)
