+++ backend/src/main/kotlin/backend/models/pizza/Cheese.kt
@@ -1,4 +1,4 @@
-package models.pizza
+package backend.models.pizza
 
 enum class Cheese(val value: Int) {
     MOZERELLA(10),
