+++ backend/src/main/kotlin/backend/models/pizza/Pizza.kt
@@ -1,7 +1,12 @@
-package models.pizza
+package backend.models.pizza
 
-data class Pizza(val name: String? = null, val sauce: Sauce, val cheese: Cheese, val toppings: List<Topping>, val id: Int? = null) {
-
+data class Pizza(
+    val name: String? = null,
+    val sauce: Sauce,
+    val cheese: Cheese,
+    val toppings: List<Topping>,
+    val id: Int? = null
+) {
 
     fun getPrice(): Int {
         return toppings.fold(0) { sum, topping -> sum + topping.value }
@@ -14,4 +19,3 @@ data class Pizza(val name: String? = null, val sauce: Sauce, val cheese: Cheese,
 
 fun main() {
 }
-
