+++ backend/src/main/kotlin/backend/models/pizza/Sauce.kt
@@ -1,4 +1,4 @@
-package models.pizza
+package backend.models.pizza
 
 enum class Sauce {
     TOMATO, BBQ, TACO;
