+++ backend/src/main/kotlin/backend/models/pizza/Topping.kt
@@ -1,4 +1,4 @@
-package models.pizza
+package backend.models.pizza
 
 enum class Topping(val value: Int) {
     MARINATED_BEEF(30),
